AIcamo/pyproject.toml
[project]
name = "cyberdefense-simulation-framework"
version = "0.1.0"
description = "High-Fidelity Simulation Framework for Advanced Cybersecurity Threat Detection and Analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "torchaudio>=2.1.0",
    "transformers>=4.35.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pywin32>=306; sys_platform == 'win32'",
    "wmi>=1.5.1; sys_platform == 'win32'",
    "psutil>=5.9.0",
    "scapy>=2.5.0",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "tqdm>=4.65.0",
    "wandb>=0.15.0",
    "scikit-learn>=1.3.0",
    "tensorboard>=2.14.0",
    "pyarrow>=14.0.0",
    "fastparquet>=2023.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
vm-automation = [
    "playwright>=1.40.0",
    "selenium>=4.15.0",
    "pywinauto>=0.6.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
